Adjust for long project titles
When project titles were long, they would push the project image up,
instead of the box down. This was undesirable. Here's how I fixed that:

* Change the `height` property on `.project-container` to `min-height`
so it could expand
* Margin on the title from the bottom to the top so it would always
push away from the image

This created a couple other visual glitches:

* Images that aren't square were being forced into that aspect ratio so
I removed the width and height property on the images
* If text happened to wrap just at the edge of the `.project-container`
it would look weird when it touched. I added padding on the left and
right of the container
* Now when titles were multiple lines, they would touch the bottom of
the container, so I added padding to the top and bottom of
`.project-container`