Use project images on homepage
I had implemented the homepage before I had project logos so I added
them to the mix. It revealed a problem though, full height logos
extended the height of the project container. To fix this, I change the
padding on the container to use `small-spacing` instead of
`base-spacing`. After looking back at my designs, this is closer to
what I had planned.