Impliment block link
Block links are when I have a block of content, like a featured project
or the logo, that needs to be linked somewhere.

This uses a technique found here:
http://stackoverflow.com/questions/796087/make-a-div-into-a-link#3494108

Inside of the link, you make a <span> tag. This span take gets
positioned absolutely and becomes the full height of the nearest
`position: relative` container. This makes the link the full size of
the container and clickable.