Fix mobile header bug
I have a link to the root of the website in the header. The link
content (Edward Loveall) is hidden, but it extends far beyond the
bounds of the logo. This was pushing the viewport on small screens out
so everything looked like it was pushed on the left side of the screen.

No good.

This adds the `overflow: hidden` property to the header so that any
thing that extends outside of it's bounds gets chopped off, visually.