Add Navigation class
This displays the navigation links that I want on any particular page.
The dynamic navigation thing started to become a huge pain, and even
when it worked perfectly, it never felt that great. This simplifies it
into two links: the links for the pages that you're not on. We'll see
how this plays out for a while.