JavaScript and Styles for audio player
Some considerations:

First, on a very short song, the player progress would stop before the
end. This adds a listener to the `ended` event so that no matter where
the progress is, when the song ends, it snaps to 100%.

Next, there is a left padding on the scrubber container. This caused the
progress div, when it was 100%, to extend past the width of the scrubber
by the padding amount. To fix this, I put the progress div inside of the
track div, because it is the width that I want. Then I removed the
`position: absolute` attribute from the progress div, and now it fits
inside the track div.