Use tabular number styles for song time
An OpenType feature allows the use of tabular (fixed width) numbers. This is so when the song's time is counting down (or changing at all) it doesn't shift left and right as the numbers get thinner or thicker, just like a monospace font.

Here's a playground where you can try out all sorts of features: http://clagnut.com/sandbox/css3/