Fix song description overflow
When the number of lines in a song's description exceded the height of
the song's title, the extra lines would appear below the player. The
whole layout was a bit hacky anyway and was improved with a rewrite.

I believe the problem was with using neat for float-based layouts. To
fix it, I converted the song layout to use flexbox instead.