Add ogg to song
Note the new initializers/paperclip.rb file. This adds a custom mime
type of application/ogg to map to .ogg files. I added this because for
whatever reason when I created a song via a test (specifically the
.by_position test in spec/models/song_spec.rb) I would get an error that
"Ogg has contents that are not what they are reported to be". This lets
the file through.