Ignore content-type for Ogg files on a Song
I don't want to remove this validation, but it was impossible to test.

When I had the content-type set to audio/ogg, it would tell me that the file "was invalid". I removed the content-type validation and uploaded the file to see what it thought the content type was. It reported application/octet-stream, which is, as far as I can tell, a generic series of bytes. Unhelpful, but whatever.

If I change the validation to application/octet-stream, uploading worked. However, when uploading the file via a test, I got the same "is invalid" error. I again removed the validation to see what the test thought the content-type was: audio/ogg.

The validation is now removed.