Lessen header font size
In a move surprising mostly me, the font size for headers (project
titles, nav titles, song titles) is now a bit smaller. This seems to fit
the text that I'm working with quite a bit better. It may be a bad move
in the long run, but I can always change it back.