Create environment specific puma configs
Puma can be configured to use a specific configuration based on the
RACK_ENV. I switched to this so I no longer needed to use a conditional
in my config and could further customize the behavior for each
environment.

Differences so far are the name of the environment (which is a bit
self-selecting as it will only load the file if that environment is
already set), deamonization for production and staging, and logging. In
prod/staging i want it to go to a file, in development, I want it to go
to STDOUT.

See here for more information:
https://github.com/puma/puma#configuration-file