Fix project layout with long project URLS
The problem was with URLs starting on the next line instead of along
side the role. I believe this is was because with a flex-box layout, the
flex-basis property determines the start size. Then it continues to grow
or shrink along the main axis (row or x axis in this case) until it
either is the size of it needs to be for its content, or it runs out of
space. In this particular case, running out of space wasn't so much of
an issue, because you can also tell the container of the flex items to
wrap instead of squish.

So the role and the website would start at 50%, which is almost always
enough for the role, but often not for the URL. So role would stay at
50% because it was satisfied and URL would grow until it fit, pushing it
to the next line.

The way I solved this was to change the flex-basis to 10%. That way both
items would grow until they were satisfied. In practice this means that
role and URL will grow just to where they need to be. Then will both
grow equally to fill up the rest of the line.

There's still a chance they will grow in such a way that URL will flow
the next line, but even that doesn't look broken.