Add browser detector
This simplifies each browser into a small string made up of the browser
name's first letter, and the major version number. For example:

Safari 9.1 -> s9
Internet Explorer 11.0 -> i11

This string is then added as a class to the body element so I can make
style changes based on the browser.