Fix sticky footer for ie 11
By default IE 11's default for flex-shrink is 0 instead of 1. Setting
this to 1 explicitly (via the combined flex attribute, like: flex: 0 1
auto) lets flexbox layout not collapse incorrectly.

One weird thing though, I had to set the body height to 1vh to make the
footer stick to the bottom of the page. No idea why this needed to be
done.