Add basic micropost model
This is the start of a microblog. A microblog is made up of many
microposts. I'm not in love with the name, but this is after trying
[snippets](http://snippets.today) and considering Tweets. Both of those
options seemed somewhat proprietary.

The goal will to have an RSS feed to view all microposts, and an API
interface to create microposts through POST requests.

Eventually, I'd love to also have the web pages for both the index and
each micropost, just like the blog.

Undecided at this point if I want to have a subdomain...

I'm basing this off of Manton Reece's microblog rss feed:
http://www.manton.org/index.php?feed=rss2&cat=847