Add guid to microposts
This is very similar to the post's guid but uses a timestamp instead of
a slug because microposts have no slug. The guids for microposts look
like:

'com.edwardloveall.microblog.1479231376'

As I said in the post guid commit (c1f2de1):

This hierarchical notation works well for apps, so I thought it may work
well for post guids. Each post's slug is both required and unique.