Add pretty_datetime helper
This formats datetimes like June 10th, 2016 - 1:02 pm. The helper uses
string interpolation for the format string because of hours in the day.
There are two formats built into strftime for 12 hour formats:

* %I - Hour of the day, 12-hour clock, zero-padded (01..12)
* %l - Hour of the day, 12-hour clock, blank-padded ( 1..12)

I don't like either because I don't like padding the hour with
anything, a zero or space. So I grab the hour directly from the time
instead.