Styles for microblog
The big change here is setting up a new layout and styles for the
microblog. There are still styles that need to be added, like
pagination, but this is a solid start.

Also, I moved the logo html into its own file for an experiment that
didn't end up working out. I think I like it better separated though.