Add ability to render markdown as plain text
This is for eventually testing if the markdown in microposts is a valid
length. For example, when renering [link](https://aol.com), that's 23
characters long, but it will only display 4 characters: "link".

Redcarpet provides a plain text renderer, but it attempts to preserve
most data, including things like link urls. The link above normally
converts into "link (https://aol.com)", so some method overriding was
necessary.