Adjust scrollDown.js to be smarter
Instead of always scrolling down when there's not enough room, the page
only scrolls if there is there is A) not enough room and B) a referrer.
The idea behind the referrer is if someone is visiting the page from
some other location, they want to see the content of that page, not the
header.