Add authentication
This uses the rails method authenticate_with_http_token at its core
which parses the HTTP header "Authorization" with a token in the form
of "Token token=12345".

authenticate_with_http_token parses the token from the headers. I
return a true/false value if it matches/doesn't match the token. Based
on that value, the app goes forward with the request or display an
error accordingly.