Fix json-feed tests
I changed `as: :json` to be part of the actual url
`microposts_feed_url(format: :json)`

I'm not sure why I had them set up like this in the first place. I
think I might have thinking that `as: :json` was sending a header that
Rails would automatically detect? I'm not sure. Either way, this works
and I have an auto-discovery url for it so feed readers should be able
to use it no problem. It defaults to RSS still.