Add rel="me" links for IndieAuth
IndieAuth is a way to use your own website as an authentication
service, as opposed to signing in with Twitter or Facebook. It works by
adding link's with a rel="me" to profiles on current services (or a PGP
key) that link back to you.

I added a couple services to test this out.

Full setup instructions here: https://indieauth.com/setup