Move dotenv-rails to all gems
When deploying, rails needs a bunch of environment variables to be set.
dotenv is usually the thing taking care of that locally and something
like Heroku would be setting environment variables server-side.

However, I'm not using Heroku and setting certain environment variables
is problematic. For example, the AWS access key and secrect key could
be different for two different applications on the server.

This is what dotenv is for. It will load a file in the root of the
application called .env. This file will contain all the environment
variables that get set _before_ rails starts (with the exception of the
RAILS_ENV and maybe RACK_ENV).

This means I can separate the variables needed for each environment in
a single file. The only thing I needed to do was move the gem outside
of the development/test group so any environment could access it.