Test for auth redirect, state, and code
When the initial call to the auth endpoint comes in to the app, the
next step is to redirect to the redirect_uri provided in the GET
params, and send with it the state (also from GET params) and a code.

The auth endpoint is supposed to be a page that asks the user if it's
ok for the web app (or whatever accessed the auth endpoint in the first
place) to access our service. This would be like if you used Twitter to
sign into an app. It first makes you sign into Twitter, then says
"So-and-so app wants to access your email and location" or whatever.
That's what this step is.

If the user allows it, it then sends a request with a code. I'm not yet
sure what this code is exactly for because it's not the final token
that the requesting app will use.

The tests here test that the controller redirects to the correct url
and contains the url params required.