Redirect to original path after sign in
If a user tries to visit /admin/something but isn't logged in yet, it
redirects them to the sign in path. After sign in, the user would just
be directed to the default page after sign in.

The solution I found was grab the original path that the user was
trying to go to, and carry it along in the sign in forms all the way to
being authenticated.

If no path was detected, it just uses the default of /admin/projects/