Use session to redirect the user after sign in
This is instead of passing around a query parameter through multiple
controllers and actions.

When it gets to the final step of choosing where to redirect users, the
session parameter is cleared. If another path is specified before the
user has logged in, it will replace the currently stored original path.
This seems better as it uses less code and less files need to know
about the original path the user was trying to visit.