Test and implement the flow for the auth endpoint
The flow is that some app requests an authorization by sending a GET
request to the authorization endpoint (/auth/new) with a bunch of
params in the url.

If the user isn't signed in, they must sign in. After sign in, they are
taken back to the /auth/new path.

This shows which app is requesting permission to act as them, and what
scope they are asking for.

The user can authorize the app. When they do, it creates an
Authorization record with the client_id (app), scope, and stores a
generated code.

If this is successful, it redirects back to a website with the code and
a state parameter. The state parameter is from the original request to
/auth/new. The website to redirect to is also from the original
request. The resulting URL ends up looking like this:

https://service.com/callback?code=abcdef&state=123456

Again, the code is generated by us, the state and base url are from the
original authorization request.