Add a token endpoint that responds to JSON and form encoded requests
The service that the user is attempting to log into will send a post
request to this website's token endpoint. That request will include:

* a `code` parameter that matches the code given from the initial
Authorization request
* a `client_id` parameter that matches the client ID from the initial
Authorization request
* a `me` parameter that identifies the user again
* a `redirect_uri` parameter that matches the redirect uri given from
the initial Authorization request

It's a lot of repeat data, but to a different endpoint!

This data can be sent via JSON:

```
{
  code: auth.code,
  me: me,
  redirect_uri: "#{auth.client_id}/callback",
  client_id: auth.client_id
}
```

or as form encoded data:

```
code=abcdef&me=https://example.com&redirect_uri=https://service.com/call
back&client_id=https://service.com
```

Either way, we check the data to see if there's an existing
authorization present that matches, then we generate a token. This
token with some other data is rendered as the response. The response
type should match the request type, JSON or form encoded:

```
{
  me: "https://example.com",
  access_token: "a54f2dc223,
  scope: "create update delete undelete"
}
```

I'm not exactly sure what scope should return but I assume it should be
the same scope as the authorization.

I'm also currently generating a token expiration date, but I'm not
actually checking that yet. Also, the token itself is pretty lame.
Needs to be longer and more secure.