Test for Authorization generate_token!
somewhat confusingly, this bang method saves the record, where as the
similar generate_code! method does not.

This is because generate_code! is not meant to be called manually and
is trigger before the record is saved (right after creation).
generate_token! however is called after the Authorization is created,
at least right now. So either you call this method and then manually
save, or it can save automatically. Saving automatically made the most
sense to me.

If I ever decide that doesn't make sense, I think a goo idea is to
either make some kind of token generator class and remove this method,
or remove the bang from the method all together to encourage manual
saving.