Verify a token endpoint
This is a request to the same token endpoint, but it uses the GET verb
instead of POST. It's meant to verify an authorization token.

The way it works is the token is sent via the Authorization header,
with a Bearer label: `Authorization: Bearer <token>`

If the token exists, it responds with the client_id, me, and scope from
the authorization. This can be in either the JSON format or the form
encoded format, depending on the Accept header from the request.