Clean up micropub creation spec
What I was trying to do with the regex is figure out if location header
pointed at the micropost. A regex isn't needed for that, just an `eq`
matcher.

Also, I found a bug in that the location header needed to point toward
the full url, not just the path which is something a regex wouldn't have
noticed!