Copy attachments to activestorage from paperclip
This migration is from here:
https://github.com/thoughtbot/paperclip/blob/master/MIGRATING.md#copy-th
e-database-data-over

I had to modify it a bit from the source because prepared statements
didn't seem to work? I wonder if that's because the PG gem went to 1.0
recently. Either way, I used `exec_params` instead of `prepare` which
seemed to work fine. I also replaced the generic placeholder arguments
`?` with numbered parameters `$1`, `$2`, etc