Copy files from paperclip folders to activestorage folders
Again from here:
https://github.com/thoughtbot/paperclip/blob/master/MIGRATING.md#moving-
local-storage-files

I made this a rake task instead of a one-off ruby script. Although I
don't plan to every run this twice on the same app, it seemed to fit
well thematically with the active_storage migration tasks.

Also, I added the new location of attachments to the gitignore.