Update project logos to use active storage attachments
I had to use polymorphic_url here for the srcset options because the
image_tag helper would convert the object to a string which looked like
`#<ActiveStorage::Variant:0x00007fad3e48f330>` which it then tried to
use for the image src. Not great.

Anyway, using polymorphic_url worked out fine because it was a string
then.