Update songs to use active storage attachments
I had to use polymorphic_url here for the multiple attachments because
the audio_tag helper would convert the object to a string which looked
like `"#<ActiveStorage::Attached::One:0x00007fd0e8bb6550>"` which it
then tried to use for the source. Not great.

Anyway, using polymorphic_url worked out fine because it was a string
which worked fine