Test subdomains with ChromDriver & Capybara
The problem here was that I am trying to use and test subdomain URLs in
the app, in non JS tests, and in JS tests. When I ran the app or non-JS
tests, it worked fine. But when I started up Capybara to run JS based
tests, the host's were different. So you'd get something like
blog.example.com in the view, and blog.localhost:38405 in the test.

I'm not sure of all the moving parts here, but here's what seemed to
help:

* Setting Capybara's server_host and rspec's default_url_options[:host]
to same thing, explicitly for all feature tests
* Setting Capybara's server_port and rspec's default_url_options[:port]
to same thing, explicitly for all feature tests
* Setting Capybara's always_include_port option to true
* Setting Capybara's app_host to the same host as everything
else so that non JS tests will match the same URLs as JS ones

This is also combined with switching from capybara-webkit to
chromdriver and selenium.

For a host, you need something that will work as a host for
RSpec/Capybara tests and for rails controllers/views/etc. You also need
something that will work when a browser is booted with chromedriver and
capybara. Using lvh.me worked because it's a domain that allows
subdomains (localhost won't) and also points to localhost because
that's where the server is running.

Then you need to make sure the proper host and ports are set for
capybara and Rails/RSpec