Convert logo to z-index
Instead of using preserve-3d which is less tested, I made the logo
"rotate" by moving left and right in the sine-wave pattern and
front-to-back using z-indexes.

I had to remove the "alternating" direction for the animation because
the z-indexes use a step animation (as opposed to a curve) and stepping
one step with an "alternate" animation has some side effects that I
couldn't figure out. It was easy enough to instead change it to be
twice as long and move back and forth. I also found the nice ability to
set the initial state of the animation in the element itself and then
just specify the 50% mark in the @keyframes. That made the keyframes
much smaller and has the side-effect of making the end state the same
as the start state.