Upgrade node-forge
This resolves a CVE: https://github.com/advisories/GHSA-92xj-mqp7-vmcj

This is a sub-sub-dependency:

* webpack-dev-server
  * selfsigned
    * node-forge

To upgrade it, I had to delete whatever the top-most dependency was
that could be upgraded. Since webpack-dev-server is up-to-date, but
selfsigned had a bump which referenced the new version of node-forge, I
could update selfsigned which would cascade the dependency tree. I went
into the yarn.lock file and deleted selfsigned, then re-ran yarn. This
bumped selfsigned and also node-forge.

https://medium.com/@ayushya/upgrading-javascript-packages-deep-dependenc
ies-using-yarn-8b5983d5fb6b