Bake in magic numbers to logo
The logo is made up of a bunch of planets which have different sizes,
and orbit positions (vertically and horizontally). These are meant to
look semi-random so they are baked in as pre-calculated number. I was
doing additional calculations in addition to these numbers with _more_
magic numbers. So instead, I just did the calculations before hand and
removed the all the magic numbers outside of the random planet
positions.