Fix logo left side
When I modified the logo last time, I removed the left-side offset. All
the planets had a skewed orbit that aligned on the left like lines in a
paragraph.

The fix was to offset them again by:

Converting each orbit width to a pixel value (percentage of logo width)
Subtracting each orbit from the logo width
Dividing that difference by 2

A conversion from percentage to pixels was necessary because pixels and
percentages can't be calculated together in SASS. I could have also
figured out the values and hard coded them.