Upgrade active storage
I was having a heck of a time getting tests to pass because of a change
in Rails 6.1 that changed something about `file_fixture_path`. This is
what `file_fixture_upload` from rspec-rails uses. The workaround I
found was to use Rack::Test::UploadedFile.new instead of
fiture_file_upload, and also upgrade active_storage