Converts Post to InternalPost
This makes `Post` have a polymorphic `postable` association which right
now is only `InternalPost`. Eventually it will also include external
posts. This aims to have almost no visible effect on any users. The
only change is that admins (like, uh, me) see "Internal Post" in the
admin UI instead of "Post".

The thing that is now rendered in the view is an `InternalPost`.
However, `Post` is is the object that is sent to the view. The way it's
rendered is using Rails implicit partial rendering where you can say
`render a_thing` and it will find `things/_thing.html.erb` and render
it. In this case, the page renders `post.postable` which is an instance
of `InternalPost` so it looks in
`internal_posts/_internal_posts.html.erb`.

The same works for the RSS feed, but finds
`internal_posts/_internal_posts.rss.builder`. The feed itself also
does't render partials like normal views because it's actually a
Builder::XmlMarkup object. The view uses render to fetch the contents
of the partial and add it into the feed.

Fetching the newest posts is also interesting because the Post itself
no longer has a `created_at` timestamp. That timestamp is normalized to
the `InternalPost` or whatever external posts in the future. It uses
some custom SQL to join internal posts and then sort them by the
`created_at` timestamp.