Ensure posts aren't calling N+1 SQL queries
This was a journey. I guess I was doing this already, and I didn't know
it. But I also thought Rails didn't support `preload`ing of polymorphic
associations. So I tried to do it myself.

I used `ActiveRecord::Associations::Preloader` to preload polymorphic
models. [This blog post][ksylvest] was helpful, but out of date as the
Preloader api changed in rails 7:

Because Posts::Preloader evaluates the query, order starts to matter. I
had to move `order` earlier in the chain. If I didn't, it would query
the posts query again, removing some of the optimizations gained by
using the preloader in the first place.

Here's the code more or less:

```ruby
module Posts
  class Preloader
    def self.preload(posts)
      preloader = ActiveRecord::Associations::Preloader.new(
        records: posts,
        associations: [:postable]
      )
    end
  end
end

class Post < ApplicationRecord
  scope :with_preloaded_postables, -> {
    self.tap do |posts|
      ::Posts::Preloader.preload(posts).call
    end
  }
end
```

To test, I counted the number of sql queries expected. To count the
number of SQL queries, I used
[ActiveSupport::Notifications][instrumentation]. This is most often
used to do some measurements all the time. To do that you want the
`subscribe` method. Since I wanted temporary measurement just for one
test, so I used the subtly different `subscribed` method. It takes a
block and unsubscribes afterward.

It also uses a custom callback class to store the number of queries. It
skips counting any internal active record queries, named as `SCHEMA`.

[ksylvest]:
https://ksylvest.com/posts/2017-08-23/eager-loading-polymorphic-associat
ions-with-ruby-on-rails
[instrumentation]:
https://guides.rubyonrails.org/active_support_instrumentation.html