Setup basic system and user
This adds two roles:

* setup_system handles some root-level setup
* create_user creates a user and group to login

This include sensitive information like user passwords. These are
stored in a [vault]. The vault's password is stored in a file
`.vault_pass` that is not included in source control. Also, an
`ansible.cfg` file was added which tells Ansible where to look for that
password.

A staging inventory was setup that points to a virtual machine run by
vagrant. It can connect with the vagrant user and setup a user that is
used for throughout the rest of the forthcoming playbook

[vault]: https://docs.ansible.com/ansible/latest/user_guide/vault.html