Add ssh configuration
Originally I had the ssh role restart the ssh service as a task.
However, this caused the play to stall. Probably because I had restarted
SSH and Ansible was trying to gather info via SSH.

The solution was switch to a handler. From [the docs]

> By default, handlers run after all the tasks in a particular play
> have been completed

Perhaps Ansible knows that handlers may stop SSH and therefore just
"fire and forget". I'm not totally sure, but I know it works!

[the docs]:
https://docs.ansible.com/ansible/latest/user_guide/playbooks_handlers.ht
ml#id4