asdf role
Various asdf plugins were not working the way I needed them to. I found
a [good script] for install asdf myself and modified it a bit to be a
little easier to read.

One thing I discovered is setting the asdf environment variables for a
block of tasks. That is, you can group tasks together with a "block"
and then set environment variables for that whole block. See [Blocks]
and [Setting the remote environment] for more details.

[good script]: https://gitlab.com/kb9zzw/garage/-/blob/master/roles/asdf/tasks/RedHat.yml
[Blocks]: https://docs.ansible.com/ansible/latest/user_guide/playbooks_blocks.html
[Setting the remote environment]: https://docs.ansible.com/ansible/latest/user_guide/playbooks_environment.html