Properly report "changed" status when asdf plugin was updated
Ansible will tell you when something changes, but some when using the
`command` or `shell` modules, it doesn't always know. Sometimes you can
rely on a specific file being added or removed, but not in the case of
an asdf-plugin update.

Every time asdf updates a plugin, the output looks like this:

```
Updating ruby to master
From https://github.com/asdf-vm/asdf-ruby
   95bb530..fdff8f6  master     -> master
Already on 'master'
Your branch is up to date with 'origin/master'.
```

The `From https://...` line sometimes multiple lines after that with a
commit SHAs are only present if the plugin was updated. All other lines
exist in both cases. To detect an updated plugin, I had check if `From`
was present in the `STDOUT` of `ansible plugin-update`. If it was (i.e.
when `find` returns something other than `-1`) it was updated. If not
(i.e. when `find` returns `-1`), it wasn't updated.