Install latest version of git
I install git in the asdf package so it can fetch everything. This
isn't always the latest version, which has the zealous diff3 (zdiff3)
diffing algorithm which I use in my dotfiles. zdiff3 is included in
[git 2.35] or higher.

This installs git by first installing the [ppa apt repository], then
installs the latest version of git.

I also had to change the order of roles to set the timezone first.
Otherwise, the time was incorrect on the server and I would get errors
like:

```
Release is not valid yet
```

And finally, [ansible-lint] is not satisfied with using `state: latest` for package dependencies. I think that's a reasonable default, but I cannot figure out how to install the latest version of git otherwise.

[ppa apt repository]: https://unix.stackexchange.com/a/170831/122572
[git 2.35]: https://raw.githubusercontent.com/git/git/master/Documentation/RelNotes/2.35.0.txt
[ansible-lint]: https://github.com/ansible/ansible-lint