Setup root postgres permissions
## hba_conf

First, this sets up an [hba_conf] which is how postgres decides who
gets to connect and how. I've set it up so:

* `postgres` user must connect via a socket locally logged in as the
  `postgres` user
* other local users connecting via a socket must be logged in as their
  user
* users over IPv4 & 6 must use a password
* users with replication privilege are just trusted

## permissions

I also set the `postgres` role's password and stored that password in
the vault. To do this, I needed to add the acl package so ansible could
[act as] the `postgres` user. Otherwise I got [errors]. Then I could
use both `become` and `become_user` to set the password.

Further `become` and `become_user` example demonstrated [here]

[hba_conf]: https://www.postgresql.org/docs/14/auth-pg-hba-conf.html
[act as]:
https://docs.ansible.com/ansible-core/2.13/user_guide/become.html#risks-of-becoming-an-unprivileged-user
[errors]: https://stackoverflow.com/questions/46352173/
[here]: https://stackoverflow.com/questions/35603049/