Cleanup postgres task
[ansible-lint] caught three things to clean up about my postgres task

## risky-shell-pipe

The [risky-shell-pipe] rule has to do with how ansible tasks can fail.
For `shell` & `command` tasks, if the command they run returns a
non-zero exit code, they fail. However, if you're using a pipe in the
`shell` module, sometimes a command can fail up the chain but the rest
of the line will continue. This can produce a false positive as the
whole thing really should fail.

Setting `set -o pipefail` and using the bash or zsh shell will stop the
whole thing in its tracks if any thing fails at any point. This then
communicates the failure correctly to ansible.

See also: https://xan.manning.io/2019/03/21/ansible-lint-rule-306.html

## risky-file-permissions

When adding a file, it should specify its permissions. I wasn't
specifying permissions on the `pg_hba.conf` file and
[risky-file-permissions] was angry.

## no-changed-when

Every time I ran the postgres role, it set the postgres password. This
isn't technically a big deal, but to be a proper task it should skip
tasks that are unnecessary. [no-changed-when] recognizes this.

To fix this particular error, I attempted to log in as the postgres
user with its password set in the `PGPASSWORD` variable. It it has a
zero exit code, then the password is set correctly and nothing needs to
happen. But if it fails, we can set the password.

[ansible-lint]: https://ansible-lint.readthedocs.io/en/latest/
[risky-shell-pipe]: https://ansible-lint.readthedocs.io/en/latest/default_rules/#risky-shell-pipe
[risky-file-permissions]: https://ansible-lint.readthedocs.io/en/latest/default_rules/#risky-file-permissions
[no-changed-when]: https://ansible-lint.readthedocs.io/en/latest/default_rules/#no-changed-when