Add app_nginx role
This configures nginx for the portfolio. This includes copying
different configs depending on if https is desired.

It also restarts nginx at the end, but only if `https` is `true`.
Normally I'd reach for Ansible's [meta module], but it doesn't support
`when` conditionals. Instead I do a quick no-op `command` which ensures
nginx is restarted, but only if `https` is `true`.

[meta module]: https://docs.ansible.com/ansible/latest/collections/ansible/builtin/meta_module.html