Asset dependency clean up
## remove webpack-jquery-ui, add jquery

When I originally switched to webpacker, I needed to add jquery which
is a supporting library for my custom jquery-ui build which lets me
drag and drop to sort items on the backend. The webpack one adds tons
of depencencies, where as jquery is just one.

## remove normalize.css

I wasn't able to get normalize.css imported from the node modules, but
I also removed it and it didn't seem to break anything. So I'm going to
keep it removed, as well as remove the `normalize-rails` gem.

## remove esbuild-sass-plugin

Finally, I was able to remove the `esbuild-sass-plugin` node module
since I'm not using esbuild to bundle sass anymore.