Allow for a custom SSH port
This can heighten security as most casual "hackers" are looking on port
22. If you pick some other random port it's that much harder for them
to try and connect to you.

There are two ssh ports to consider: the one that ansible uses to
connect to the server, and the one that the server sets as the port for
people to connect to.

You'd think these would be the same, and technically they are. However,
when vagrant is involved, it (by default) forwards the normal ssh 22
port to 2222.

So the way I set this up is:

* Put a custom port in the vault
* Expose that port in the group_vars
* Each inventory block has two ssh port options. They can be the same
(normal) or different (in a case like vagrant with port forwarding)
  * One for ansible to use to connect
  * One for ansible to set up the remote ssh config `custom_ssh_port`
* Any playbook task can use `custom_ssh_port` for everything except
actually connecting to the machine

Configuring SSH also had to be moved to the bootstrap portion of the
playbook, becuse the remainder of the playbook uses the now-custom port
to connect for everything else. I also needed to update the UFW
firewall config to let in SSH connections on the custom port as well.