Ensure firewall is setup before user connects
Because there's the possibility of connecting to SSH on a custom port,
the firewall has to open access to that port. That means the firewall
role had to come before the deployer user connects.

To support a custom port, I had to:

* install UFW (not always present)
* open the custom port

Ansible has a module to set rules via UFW. Normally I'd set a name like
`OpenSSH` to enable ssh connections, but that (as far as I can tell)
sets the port to 22, not a custom port. So I had to remove the name and
just specify the port.

It seems that by default, ubuntu opens port 22 for SSH connections so I
hadn't noticed any of this before.

## Debugging info

I had to debug this for a bit because I wasn't able to access the
server on the custom port. Wiping the server and reinstalling fixed it.
I think what broke it is me running `iptables -F` which flushes
(resets) the firewall (iptables) configuration, but doesn't remove all
the custom "chains" setup by UFW (more on this below). Here are some
notes I made while learning about how iptables and ufw work:

The way iptables works is it has multiple "chains" that look at traffic
and decide what to do with it; accept it, drop it, move it, etc. The
chain we care about in this case is the `INPUT` chain which processes
incoming traffic. Chains contain rules and can call each other to help
organize rules.

UFW creates custom chains which then need to be referenced by the
default iptables chains to process the UFW rules. UFW puts custom rules
in the `ufw-user-input` chain. iptables was not referencing that chain,
so it would not evaluate any rules.

You can manually add the chain like so:

```sh
sudo iptables -A INPUT -p tcp --dport 0000 -j ufw-user-input
```

To test rules on the `INPUT` chain, you can use netcat from a different
server and see if it accepts the traffic:

```
nc -zv <server ip> <port to test>
```