Ensure UFW is started after netfilter-persistent.service
When I restarted the server, the UFW chains would not be present in
iptables config. This is what happened in
2ae1a4cfedd86a783b6496786c40f6b730d72018, not me flushing the tables.
The problem that I figured out from [this blog post] is the service
needs to be started after the `netfilter-persistent.service`.
`netfilter-persistent` (or its old name `iptables-persistent`) writes
and retrieves iptables rules so they can be saved when new rules are
added and restored after restart. UFW needs to be loaded after this or
the custom UFW chains will not be restored meaning no custom firewall
rules will work.

[this blog post]:
https://devtidbits.com/2019/07/31/ufw-service-not-loading-after-a-reboot
/